feat: 어드민 회원 목록 API 연결 (#22)
* style: 레이아웃 관련 아이콘 생성

* feat: MainLayout 구현 및 관련 컴포넌트 구현

* feat: 사이드바 메뉴 객체 작성

* chore: svgr/webpack 설치 및 public alias 작성

* design: 디자인 수정

* feat: 불필요한 코드 삭제, 인포 팝업 구현, useAccount에 role 추가

* feat: sidebar store 구현

* fix: 디자인 수정

* fix: InfoPopup 수정

* feat: 사이드바와 페이지 연결

* fix: 회원가입 버튼 디자인 수정

* feat: 매장 정보 페이지 UI 구현

* feat: 모바일 사이드바 네비게이션 연결

* feat: 원산지 항목 추가 구현

* feat: 버튼 기본 스타일 추가

* feat: 버튼 기본 스타일 추가

* feat: submit 버튼 위치 변경

* style: 불필요한 주석 삭제

* feat: useSidebar 연결

* fix: app page에서 메뉴 항목 처리

* feat: 이메일 인증 또는 매장 등록을 하지 않은 사이드바 구현

* feat: 메뉴 및 메뉴 항목 로컬 저장

* fix: 메뉴 아이콘 처리 수정

* fix: account 관련 수정

* fix: 아이콘에 width, height 삭제

* feat: Table 컴포넌트 분리

* fix: 메뉴 오타 수정

* feat: Table 컴포넌트 컴파운드로 변경

* fix: 매장 등록 승인 타이틀 수정

* feat: 모바일 테이블에 신청일 추가

* feat: 매장 검색 인풋 구현

* rename: 파일명 변경

* fix: 기존 api 삭제

* style: 주석 삭제

* feat: 승인된 매장이 없을 경우 페이지 수정

* fix: 토큰 없을 때 에러 표시 삭제

* fix: MainLayout 수정

* feat: stores API 복구

* style: 불필요한 코드 삭제

* design: 디자인 수정

* feat: 잘못된 API 연결 삭제

* fix: 페이지네이션 수정

* fix: 빌드 에러 수정

* fix: 빌드 에러 수정

* refactor: MainLayout 빌드 에러 해결

* feat: layout에 suspense 추가

* fix: 서버 컴포넌트에서 ssr: false 삭제

* fix: 매장 페이지 height 수정

* fix: submit에 disabled 처리 추가

* feat: 사이드바에 다중 매장 처리

* feat: admin menu 처리

* feat: admin menu 처리

* refactor: 초기 렌더링 이미지 priority 적용

* fix: 사이드바 메뉴 처리 수정

* fix: SearchBar 분리

* fix: Button 컴포넌트 수정

* feat: Dropdown 구현

* feat: 어드민 회원 관리 레이아웃 구현

* feat: Table 컴포넌트 수정

* fix: 다지안 및 스타일 변경

* fix: table row border 처리

* fix: mobile - 회원 관리 수정

* feat: 스토어 테이블에 모달 연결

* feat: 회원 정보 모달 구현

* fix: 검색바 margin 수정

* feat: useDebounce 구현

* feat: 어드민 회원 관리 API 연결

* feat: status 별 컬러 데이터 분리

* feat: interceptor 수정

* style: lint 규칙에 따라 코드 수정 및 디자인 코드 수정

* style: bigint 대소문자 오타 수정

* fix: 에러 해결

* fix: 디바운스 검색어 관련 오류 해결

* fix: 날짜 처리 분리

* style: 오타 수정

* fix: 테이블 수정

* design: 디자인 수정

* feat: 모달에 QueryProviders 추가

* fix: 매장 여부 드롭다운 데이터 수정

* feat: UserInfoModal inner height 수정

* style: 스타일 수정

* fix: 버튼 수정

+++ src/lib/axios/instance.ts
@@ -55,6 +55,7 @@ const setupInterceptors = (axiosInstance: AxiosInstance) => {
     (response) => response,
     // 2xx 외의 범위 상태 코드(실패)
     async (error) => {
+      console.log(error);
       const originalRequest = error.config;
       if (error.response.status === 401 && !originalRequest.isRetry) {
         originalRequest.isRetry = true;
@@ -74,10 +75,18 @@ const setupInterceptors = (axiosInstance: AxiosInstance) => {
             return await instance(originalRequest);
           }
         } catch (refreshError) {
-          console.error("토큰 갱신 실패:", refreshError);
-          await deleteCookie("accessToken");
-          await deleteCookie("refreshToken");
-          return Promise.reject(refreshError);
+          if (
+            ["FORBIDDEN", "UNAUTHORIZED"].includes(
+              (refreshError as any).response.data.code
+            )
+          ) {
+            alert((refreshError as any).response.data.message);
+            window.location.href = "/login";
+          } else {
+            await deleteCookie("accessToken");
+            await deleteCookie("refreshToken");
+            return Promise.reject(refreshError);
+          }
         }
       }
       if (error.response.status === 403) {
